Provided both convolve and fftconvolve options for all objects.
DataArray._fftconvolve() has been renamed as DataArray._convolve().
This now takes a function argument that can be either
scipy.signal.convolve or scipy.signal.fftconvolve, depending on
whether one wants to use the discrete convolution equation or the
Fourier convolution theorem to perform the convolution.

Spectrum,Image and Cube.fftconvolve() have all been modified to call
DataArray._convolve() with the function argument set to
scipy.signal.fftconvolve().

Spectrum,Image and Cube.convolve() functions have now been created,
which call DataArray._convolve() with its function argument set to
scipy.signal.convolve().

Note that it is useful to have the option of both fftconvolve and
convolve, because the latter is more efficient for small convolution
kernels, whereas the latter is more efficient for larger convolution
kernels. It isn't possible to automatically decide which algorithm to
use, because the correct choice often depends on whether the user
cares about speed or memory usage.