In Image.align_with_image() start by extracting the sub-image to be aligned.
The align_with_image() function rotates then resamples one image onto
the coordinate grid of another. To improve the efficiency of the
rotation and resampling phases of this operation, this patch precedes
these operations by truncating the image to a rectangular area that
just encloses the region that needs to be rotated and resampled.

This greatly reduces the resource usage of this function when it is
used to resample part of a large HST UDF image onto the coordinate
grid of a MUSE UDF image. For example, on musicos1 it reduced the
duration of one call to this function by a factor of 7, from a
duration of 3 minutes, to a duration of 26 seconds. It also reduced
the peak memory usage by a factor of 5, from 15GB to 3GB.