In Cube.bandpass_image(), add filter truncation and interpolation argument.
If their isn't a complete overlap between the bandpasses of the
filter-curve and the cube, truncate the filter at the edges of the
cube instead of raising an exception. When this happens, display a
warning (using the logger) which tells the user what fraction of the
bandpass that has been truncated.

When integrating the filter curve over each wavelength channel of the
cube, use linear interpolation by default, rather than cubic. This
greatly speeds up the function, and it makes no noticeable difference
to the final image when used with HST ACS filter curves. An optional
argument has also been added to give the user the option of selecting
cubic interpolation.