Improved the bandpass truncation warning in Cube.bandpass_image().
The original truncation warning simply indicated what fraction of the
wavelength coverage of the filter bandpass lay outside the wavelength
coverage of the cube. That wasn't useful because it gave no idea of
how much flux was being lost. The new warning reports the fraction of
the integrated filter curve that is being truncated, and thus
indicates how significant the truncation is.