Restore single-pixel dimensions that numpy removes in DataArray.__getitem__().
In numpy, if a slice specification includes int's for one or more
dimensions, those dimensions are removed from the shape. For example,
Cube.data[1:10,2,3:8], ends up with a 2D shape of [2,5], whereas in
DataArray.__getitem__() we need it to return a 3D sub-cube of shape
[2,1,5]. To resolve this, reshape the data, variance and mask arrays,
where needed, to reinstate the removed dimensions.