Make sure that we have enough pixels with a flux in each slice.
Fix a bug reported by J. Zabl where a slice can have a few points but
not enough to have at least one pixel with a computed flux. And add
a small optimization to avoid computing the mad sigma clip when there is
not enough points (as it is not used later in this case).