Fix reading of FITS table with Astropy 3.0
Astropy 3.0 serialize .format and .description in yaml comments, which
causes the table to not be readable by Catalog.read (inside Astropy it
is then read as QTable which is not parent class from Catalog).  So for
now we just remove .format and .description in this case. Should be
fixed in 3.0.1
https://github.com/astropy/astropy/issues/7181