[get_image] Remove background before averaging / summing
In mpdaf.obj.Cube.get_image, remove the background from the cube before
computing the sum or the mean.  It will make it easier to add new
aggregation functions.

Also the background is now the average of the mean in the below and
above regions.  Before, when using is_sum, the method was equivalent to
taking the mean of all the pixels below and above.  The new method may
give slightly different results when the two region are not made with
the same number of pixels, at the edges of the spectrum.