Fix pickle of MPDAF objects when the header does not match the data
When pickling an MPDAF object like a spectrum or image extracted from
a cube, the header is not updated. But we need a correct header to
recreate the objects when unpickling. The current method was not always
working e.g. when WCSAXES is missing, hence try to fix this by creating
an HDU which should contain valid keywords.